HouseKeeping: Redis & MemCached implementation refactoring
(cherry picked from commit 90d672ce2f6271378840af1673d7e0e653a0f81a)
