+++ gxcache-redis/src/test/java/com/genexus/test/cache/redis/TestRedisCacheClient.java
@@ -1,5 +1,6 @@
-package com.genexus.cache.redis;
+package com.genexus.test.cache.redis;
 
+import com.genexus.cache.redis.RedisClient;
 import com.genexus.db.CacheValue;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -8,7 +9,6 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import redis.clients.jedis.Jedis;
-import redis.clients.jedis.JedisPool;
 
 import java.net.URISyntaxException;
 import java.util.concurrent.ThreadLocalRandom;
